refactor: consolidate embed xmp module under plugin
