+++ pkgs/plugins/EmbedXMP/pyproject.toml
@@ -18,7 +18,6 @@ authors = [{ name = "Jacob Stewart", email = "jacob@swarmauri.com" }]
 dependencies = [
     "swarmauri_core",
     "swarmauri_base",
-    "swarmauri_embed_xmp",
 ]
 
 [project.urls]
@@ -27,7 +26,6 @@ Homepage = "https://github.com/swarmauri/swarmauri-sdk"
 [tool.uv.sources]
 swarmauri_core = { workspace = true }
 swarmauri_base = { workspace = true }
-swarmauri_embed_xmp = { workspace = true }
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
@@ -40,4 +38,6 @@ dev = [
 ]
 
 [tool.poetry]
-packages = [{ include = "EmbedXMP" }]
+packages = [
+    { include = "EmbedXMP" },
+]
