+++ pkgs/pyproject.toml
@@ -161,7 +161,6 @@ members = [
     "standards/swarmauri_signing_ssh",
     "standards/swarmauri_certs_x509verify",
     "plugins/EmbedXMP",
-    "standards/swarmauri_embed_xmp",
     "standards/swarmauri_xmp_gif",
     "standards/swarmauri_xmp_jpeg",
     "standards/swarmauri_xmp_mp4",
