--- pkgs/standards/swarmauri_embed_xmp/LICENSE
@@ -1,2 +1,0 @@
-This project is licensed under the Apache License 2.0.
-See the root LICENSE file of the swarmauri-sdk repository.
