--- pkgs/standards/swarmauri_embed_xmp/pyproject.toml
@@ -1,38 +1,0 @@
-[project]
-name = "swarmauri_embed_xmp"
-version = "0.1.0"
-description = "Dynamic manager that discovers Swarmauri XMP handlers at runtime."
-license = "Apache-2.0"
-readme = "README.md"
-requires-python = ">=3.10,<3.13"
-classifiers = [
-    "License :: OSI Approved :: Apache Software License",
-    "Natural Language :: English",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.12",
-    "Programming Language :: Python :: 3.13",
-    "Development Status :: 3 - Alpha",
-]
-authors = [{ name = "Jacob Stewart", email = "jacob@swarmauri.com" }]
-dependencies = [
-    "swarmauri_core",
-    "swarmauri_base",
-]
-
-[project.urls]
-Homepage = "https://github.com/swarmauri/swarmauri-sdk"
-
-[tool.uv.sources]
-swarmauri_core = { workspace = true }
-swarmauri_base = { workspace = true }
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
-
-[dependency-groups]
-dev = [
-    "pytest>=8.0",
-    "ruff>=0.9",
-]
